WIP PR fix: adapting s3, minio and oss unit tests to the new pattern
