+++ storage/minio/package.json
@@ -22,7 +22,10 @@
     "lint": "eslint ./src/**/*.ts 1>&2",
     "lint-fix": "eslint --fix ./src/**/*.ts 1>&2 && sort-package-json",
     "spell-check": "cspell \"**\" --config ./node_modules/@itwin/object-storage-common-config/cspell.json",
-    "test:unit": "mocha lib/test/unit/**/*.test.js --color --exit",
+    "test:unit": "npm run test:unit:backend",
+    "test:unit:backend": "npm run test:unit:backend:common && npm run test:unit:backend:local",
+    "test:unit:backend:local": "mocha lib/test/unit/backend/**/*.test.js --color --exit",
+    "test:unit:backend:common": "node lib/test/unit/backend/RunCommonUnitTests.js",
     "test:integration:backend": "npm run test:integration:downloadMinio && npm-run-all -p -r test:integration:startMinio test:integration:backend:runTests",
     "test:integration:backend:runTests": "wait-on tcp:9000 -t 30000 && node lib/test/integration/backend/RunIntegrationTests.js",
     "test:integration:frontend": "npm run test:integration:downloadMinio && npm-run-all -p -r test:integration:startMinio test:integration:frontend:startServerRunTests ",
@@ -49,6 +52,7 @@
     "@itwin/object-storage-tests-backend": "workspace:*",
     "@itwin/object-storage-tests-frontend": "workspace:*",
     "@itwin/object-storage-tests-unit": "workspace:*",
+    "@itwin/object-storage-tests-backend-unit": "workspace:*",
     "@types/chai": "^4.1.7",
     "@types/chai-as-promised": "^7.1.2",
     "@types/minio": "~7.0.11",
