+++ storage/minio/src/test/unit/backend/MinioClientStorageBindings.test.ts
@@ -4,32 +4,23 @@
  *--------------------------------------------------------------------------------------------*/
 import { Container } from "inversify";
 
-import { ClientStorage, Types as CoreTypes } from "@itwin/object-storage-core";
+import { ClientStorage } from "@itwin/object-storage-core";
 import {
   DependencyBindingsTestCase,
   testBindings,
 } from "@itwin/object-storage-tests-unit";
 
-import { S3ClientStorage, S3ClientStorageBindings } from "../client";
-import { S3ClientWrapperFactory } from "../frontend";
+import { MinioClientStorage, MinioClientStorageBindings } from "../../../client";
 
-describe(`${S3ClientStorageBindings.name}`, () => {
-  const clientBindings = new S3ClientStorageBindings();
+describe(`${MinioClientStorageBindings.name}`, () => {
+  const clientBindings = new MinioClientStorageBindings();
 
   describe(`${clientBindings.register.name}()`, () => {
     const bindingsTestCases: DependencyBindingsTestCase[] = [
       {
-        testedClassIdentifier: CoreTypes.Client.clientWrapperFactory.toString(),
-        testedFunction: (container: Container) =>
-          container.get<S3ClientWrapperFactory>(
-            CoreTypes.Client.clientWrapperFactory
-          ),
-        expectedCtor: S3ClientWrapperFactory,
-      },
-      {
         testedClassIdentifier: ClientStorage.name,
         testedFunction: (container: Container) => container.get(ClientStorage),
-        expectedCtor: S3ClientStorage,
+        expectedCtor: MinioClientStorage,
       },
     ];
     testBindings(clientBindings, undefined, bindingsTestCases);
