+++ storage/minio/src/test/unit/backend/RunCommonUnitTests.ts
@@ -1,0 +1,31 @@
+/*---------------------------------------------------------------------------------------------
+ * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
+ * See LICENSE.md in the project root for license terms and full copyright notice.
+ *--------------------------------------------------------------------------------------------*/
+import "reflect-metadata";
+
+import { StorageUnitTests } from "@itwin/object-storage-tests-backend-unit";
+import { createStubInstance } from "sinon";
+import { MinioClientStorage } from "../../../client";
+import { S3ClientWrapper, S3ClientWrapperFactory } from "@itwin/object-storage-s3";
+import { MinioServerStorage } from "../../../server";
+import { mockPresignedUrlProvider, mockTransferConfigProvider } from "@itwin/object-storage-tests-unit";
+
+const mockS3ClientWrapper = createStubInstance(S3ClientWrapper);
+const serverStorage = new MinioServerStorage(
+  mockS3ClientWrapper,
+  mockPresignedUrlProvider,
+  mockTransferConfigProvider
+);
+
+const mockS3ClientWrapperFactory = createStubInstance(S3ClientWrapperFactory);
+const clientStorage = new MinioClientStorage(mockS3ClientWrapperFactory);
+
+const tests = new StorageUnitTests(
+  serverStorage,
+  clientStorage,
+);
+tests.start().catch((err) => {
+  process.exitCode = 1;
+  throw err;
+});
