+++ storage/oss/src/test/unit/backend/OssServerStorageBindings.test.ts
@@ -15,7 +15,7 @@ import {
 import {
   OssServerStorageBindings,
   OssTransferConfigProvider,
-} from "../../server";
+} from "../../../server";
 
 describe(`${OssServerStorageBindings.name}`, () => {
   const serverBindings = new OssServerStorageBindings();
